Make NaN check an APEBench utility
