+++ apebench/__init__.py
@@ -17,6 +17,7 @@ from ._run import (
 )
 from ._utils import (
     aggregate_gmean,
+    check_for_nan,
     melt_data,
     melt_loss,
     melt_metrics,
@@ -51,4 +52,5 @@ __all__ = [
     "relative_by_config",
     "split_train",
     "components",
+    "check_for_nan",
 ]
